requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0
opencv-python>=4.8.0
faiss-cpu>=1.7.4
numpy>=1.24.0
scikit-learn>=1.3.0